Move scoring and LMR tweaks.
Only do killer move lookup on non-cap/promo moves.

If not a killer move provide a minor score boost if it has a positive
history value.

Don't apply LMR to king moves.

Loosen the reigns a little and allow LMR on moves with a history score
of 0 - which could be a move that has never been searched, but the
chances of this being the case are very low.