Remove NMP mate threat extensions. Add delta pruning.
My (admittedly limited testing) didn't show any significant difference
when NMP mate threat extensions were enabled, so that feature has been
removed.

This delta pruning implementation only uses a delta of 50 because it's
using SEE value instead of cap value for comparison.  SEE value should
be roughly the same as doing an actual quiescence search in the vast
majority of cases, so we should be able to get away with a smaller error
margin.