Testing more aggressive LMR. Also expand aspiration window faster.
Late move reductions of 1 results in large positional evaluation changes
as compared to reduction of 2.  Doing a default reduction of 2
stabilizes the scoring a bit which I'm guessing reduces the number of
re-searches due to LMR nodes returning alpha.

Initial testing doesn't appear to surface acute tactical weakness
compared to reduction of 1 - the opposite seems to be occurring
actually.  My guess is this is because aggressive LMR ends up choosing
lines with fewer obvious options for the defending side (which
translates into tempo).  It also tends to stick to attacking moves that
are more on the obvious/sound side rather than awkward manoeuvres.