Added razoring with default delta of 300
I've never liked razoring, it's effectiveness is usually a wash in my
experience.  What you gain in search depth is easily countered by loss
in tactical strength.

But, as long as it's configurable (so it can be disabled or set to a
very high delta) it doesn't hurt to add it.