Changing RZR default back to 0.
Results with RZR set to 80 were terrible.  Results with static NMP at
depth == 1 varied wildly, but at least it averaged out to be about the
same as not having that feature enabled.  RZR with delta 80 produces
abysmal results all around.

In short: still not a fan of razoring.  A high razor delta doesn't
provide enough benefit to compensate for the tactical degradation.  And
a low delta is just plain bad.